Fix ProfileProvider.credentials when ~/.aws/config doesn't exist
